vfs::dir remove GType
The were a lot of problems caused by vfs::dir being a GType. The
vfs::dir cache would not work or segfault when using cpp types and had
to use char* as its key. Cache lookup could fail even if a vfs::dir for
that path existed. vfs::dir::path could not be a std::filesystem::path,
but it could be a std::string. There were more problems but those were
some of the main ones.

The vfs::dir cache is disabled in this commit. There are signal connect
problems with multiple tabs having the same vfs::dir. This has already
been fixed and the cache will be reenabled after the signal patches
get pushed.